trends/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.11
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
kombu==5.4.2
lxml==5.3.1
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.3
pandas==2.2.3
prompt_toolkit==3.0.50
propcache==0.3.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2025.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
